tune-gpt/requirements.txt
openai
python-dotenv
# doordash api
pyjwt
scrapy
loguru
langchain
# Developer tools
black[jupyter]
pynvim
